+++ src/client/power-pages/commonUtility.ts
@@ -69,64 +69,7 @@ export function getDeletePathUris(fsPath: string,
     return pathUris;
 }
 
-export async function updateEntityPathNames(oldUri: vscode.Uri,
-    newUri: vscode.Uri,
-    oldFileProperties: IFileProperties,
-    fileEntityType: PowerPagesEntityType
-) {
-    const entityFolderName = getEntityFolderName(oldUri.fsPath);
-    if (isValidUri(oldUri.fsPath) && oldFileProperties.fileName && isValidRenamedFile(oldUri.fsPath, entityFolderName, oldFileProperties.fileName, fileEntityType)) {
-        const newFileProperties = getFileProperties(newUri.fsPath);
-        const entityFolderIndex = getEntityFolderPathIndex(newUri.fsPath, oldFileProperties.fileName, fileEntityType);
-        const fileEntityFolderPath = oldFileProperties.fileFolderPath.substring(0, entityFolderIndex);
-        const allFilesInFolder = await vscode.workspace.findFiles(fileEntityType === PowerPagesEntityType.WEBFILES ?
-            `**/${entityFolderName}/${oldFileProperties.fileName}.*` : `**/${entityFolderName}/${oldFileProperties.fileName.toLowerCase()}/**/*`);
-
-        if (fileEntityType !== PowerPagesEntityType.ADVANCED_FORMS) {
-            await Promise.all(allFilesInFolder.map(async file => {
-                const fileProperties = getFileProperties(file.fsPath);
-                if (fileProperties.fileCompleteName) {
-                    const newFilePath = file.fsPath.replace(fileProperties.fileCompleteName, [newFileProperties.fileName, fileProperties.fileExtension].join('.'));
-                    if (newFilePath !== file.fsPath) {
-                        await vscode.workspace.fs.rename(file, vscode.Uri.file(newFilePath), { overwrite: true });
-                    }
-                }
-            }));
-        }
-
-        // FolderName update
-        if (newFileProperties.fileName && fileEntityType !== PowerPagesEntityType.WEBFILES) {
-            await vscode.workspace.fs.rename(vscode.Uri.file(fileEntityFolderPath),
-                getUpdatedFolderPath(fileEntityFolderPath, oldFileProperties.fileName, newFileProperties.fileName),
-                { overwrite: true });
-        }
-    }
-}
-
-export async function fileRenameValidation(oldUri: vscode.Uri,
-    newUri: vscode.Uri,
-    oldFileProperties: IFileProperties
-) {
-    let success = true;
-    if (isValidUri(oldUri.fsPath) && oldFileProperties.fileName) {
-        const newFileProperties = getFileProperties(newUri.fsPath);
-
-        if (oldFileProperties.fileExtension !== newFileProperties.fileExtension && newFileProperties.fileName && oldFileProperties.fileExtension) {
-            await vscode.workspace.fs.rename(newUri, getValidatedEntityPath(newFileProperties.fileFolderPath, newFileProperties.fileName, oldFileProperties.fileExtension), { overwrite: true });
-        } else if (oldFileProperties.fileExtension === '') { // revert name changes - entity folder re-naming is not a valid scenario
-            await vscode.workspace.fs.rename(newUri, oldUri, { overwrite: true });
-            success = false;
-        }
-    }
-    return success;
-}
-
-export function getCurrentWorkspaceURI(fsPath: string): vscode.Uri | undefined {
-    const workspaceFolder = vscode.workspace.getWorkspaceFolder(vscode.Uri.file(fsPath));
-    return workspaceFolder ? workspaceFolder.uri : undefined;
-}
-
-function isValidUri(fsPath: string): boolean {
+export function isValidUri(fsPath: string): boolean {
     let validUri = true;
 
     EntityFolderName.forEach(folderName => {
@@ -138,7 +81,7 @@ function isValidUri(fsPath: string): boolean {
     return validUri;
 }
 
-function getEntityFolderName(fsPath: string): string {
+export function getEntityFolderName(fsPath: string): string {
     let entityFolderPath = '';
 
     EntityFolderName.forEach(folderName => {
@@ -150,20 +93,25 @@ function getEntityFolderName(fsPath: string): string {
     return entityFolderPath;
 }
 
-function getEntityFolderPathIndex(fsPath: string, fileName: string, fileEntityType: PowerPagesEntityType) {
+export function getEntityFolderPathIndex(fsPath: string, fileName: string, fileEntityType: PowerPagesEntityType) {
     return fileEntityType === PowerPagesEntityType.WEBFILES ? fsPath.indexOf(`\\${WebFilesFolder}\\`) + WebFilesFolder.length + 2 :
         fsPath.indexOf(`\\${fileName?.toLowerCase()}\\`) + fileName?.length + 2; // offset for path separator
 }
 
-function getValidatedEntityPath(folderPath: string, fileName: string, fileExtension: string): vscode.Uri {
+export function getValidatedEntityPath(folderPath: string, fileName: string, fileExtension: string): vscode.Uri {
     return vscode.Uri.file(folderPath + [fileName, fileExtension].join('.'));
 }
 
-function isValidRenamedFile(fsPath: string, entityFolderName: string, fileName: string, fileEntityType: PowerPagesEntityType): boolean {
+export function isValidRenamedFile(fsPath: string, entityFolderName: string, fileName: string, fileEntityType: PowerPagesEntityType): boolean {
     return fileEntityType === PowerPagesEntityType.WEBFILES ? fsPath.includes(`\\${entityFolderName}\\${fileName}`) :
         fsPath.includes(`\\${entityFolderName}\\${fileName.toLowerCase()}\\`);
 }
 
-function getUpdatedFolderPath(fsPath: string, oldFileName: string, newFileName: string): vscode.Uri {
+export function getUpdatedFolderPath(fsPath: string, oldFileName: string, newFileName: string): vscode.Uri {
     return vscode.Uri.file(fsPath.replace(`\\${oldFileName.toLowerCase()}\\`, `\\${newFileName.toLowerCase()}\\`));
+}
+
+export function getCurrentWorkspaceURI(fsPath: string): vscode.Uri | undefined {
+    const workspaceFolder = vscode.workspace.getWorkspaceFolder(vscode.Uri.file(fsPath));
+    return workspaceFolder ? workspaceFolder.uri : undefined;
 }
