+++ src/client/power-pages/validationDiagnostics.ts
@@ -23,7 +23,7 @@ export async function validateTextDocument(uri: vscode.Uri, patterns: RegExp[], 
                 severity: vscode.DiagnosticSeverity.Warning,
                 range: new vscode.Range(textDocument.positionAt(m.index), textDocument.positionAt(m.index + m[0].length)),
                 // localize("powerPages.searchByNameReferenceMessage", `Deleted file might be referenced by name here.`)
-                message: `${m[0]}: ` + (searchByName ? `Deleted file might be referenced by name here.` : ""),
+                message: `${m[0]}: ` + (searchByName ? `File might be referenced by name here.` : ""),
                 source: 'ex',
                 // relatedInformation: [
                 //     new vscode.DiagnosticRelatedInformation(new vscode.Location(textDocument.uri, new vscode.Range(new vscode.Position(1, 8), new vscode.Position(1, 9))), 'first assignment to `x`')
